+ Issue 157: 	Statistic correction

+++ modules/webserver/src/main/java/wisematches/server/web/controllers/playground/player/PlayerSearchController.java
@@ -29,66 +29,66 @@ import java.util.Map;
 @Controller
 @RequestMapping("/playground/players")
 public class PlayerSearchController extends AbstractSearchController<PlayerEntityBean, PlayerSearchArea> {
-	private GameMessageSource messageSource;
-	private PlayerStateManager stateManager;
+    private GameMessageSource messageSource;
+    private PlayerStateManager stateManager;
 
-	private static final List<PlayerSearchArea> AREAS = Arrays.asList(PlayerSearchArea.FRIENDS, PlayerSearchArea.FORMERLY, PlayerSearchArea.PLAYERS);
+    private static final List<PlayerSearchArea> AREAS = Arrays.asList(PlayerSearchArea.FRIENDS, PlayerSearchArea.FORMERLY, PlayerSearchArea.PLAYERS);
 
-	private static final Log log = LogFactory.getLog("wisematches.server.web.search");
+    private static final Log log = LogFactory.getLog("wisematches.server.web.search");
 
-	public PlayerSearchController() {
-		super(new String[]{"nickname", "ratingG", "ratingA", "language", "averageMoveTime", "lastMoveTime"});
-	}
+    public PlayerSearchController() {
+        super(new String[]{"nickname", "ratingG", "ratingA", "language", "averageMoveTime", "lastMoveTime"});
+    }
 
-	@RequestMapping("")
-	public String showPlayersSearchForm(@RequestParam(value = "area", required = false, defaultValue = "FRIENDS") PlayerSearchArea area, Model model) {
-		model.addAttribute("searchArea", area);
-		model.addAttribute("searchAreas", AREAS);
-		model.addAttribute("searchColumns", getColumns());
-		model.addAttribute("searchEntityDescriptor", getEntityDescriptor());
-		return "/content/playground/players/view";
-	}
+    @RequestMapping("")
+    public String showPlayersSearchForm(@RequestParam(value = "area", required = false, defaultValue = "FRIENDS") PlayerSearchArea area, Model model) {
+        model.addAttribute("searchArea", area);
+        model.addAttribute("searchAreas", AREAS);
+        model.addAttribute("searchColumns", getColumns());
+        model.addAttribute("searchEntityDescriptor", getEntityDescriptor());
+        return "/content/playground/players/view";
+    }
 
-	@ResponseBody
-	@RequestMapping("load.ajax")
-	public Map<String, Object> load(@RequestParam("area") String areaName, @RequestBody Map<String, Object> request, Locale locale) {
-		final Personality personality = getPersonality();
-		final PlayerSearchArea area = PlayerSearchArea.valueOf(areaName.toUpperCase());
-		if (log.isDebugEnabled()) {
-			log.debug("Loading players for area: " + area + " for player " + personality);
-		}
-		return loadData(personality, request, area, locale);
-	}
+    @ResponseBody
+    @RequestMapping("load.ajax")
+    public Map<String, Object> load(@RequestParam("area") String areaName, @RequestBody Map<String, Object> request, Locale locale) {
+        final Personality personality = getPersonality();
+        final PlayerSearchArea area = PlayerSearchArea.valueOf(areaName.toUpperCase());
+        if (log.isDebugEnabled()) {
+            log.debug("Loading players for area: " + area + " for player " + personality);
+        }
+        return loadData(personality, request, area, locale);
+    }
 
-	@Override
-	protected void convertEntity(PlayerEntityBean info, Personality personality, Map<String, Object> map, Locale locale) {
-		map.put("nickname", ServicePlayer.get(info.getPid(), info.getNickname(), stateManager));
-		if (info.getLanguage() != null) {
-			map.put("language", messageSource.getMessage("language." + info.getLanguage().code(), locale));
-		} else {
-			map.put("language", messageSource.getMessage("search.err.language", locale));
-		}
-		map.put("ratingG", info.getRatingG());
-		map.put("ratingA", info.getRatingA());
-		map.put("lastMoveTime", info.getLastMoveTime() != null ? messageSource.formatDate(info.getLastMoveTime(), locale) : messageSource.getMessage("search.err.nomoves", locale));
-		map.put("activeGames", info.getActiveGames());
-		map.put("finishedGames", info.getFinishedGames());
-		map.put("playerOnline", stateManager.isPlayerOnline(Personality.person(info.getPid())));
-		map.put("averageMoveTime", info.getAverageMoveTime() != 0 ? messageSource.formatMinutes(info.getAverageMoveTime() / 1000 / 60, locale) : messageSource.getMessage("search.err.nomoves", locale));
-	}
+    @Override
+    protected void convertEntity(PlayerEntityBean info, Personality personality, Map<String, Object> map, Locale locale) {
+        map.put("nickname", ServicePlayer.get(info.getPid(), info.getNickname(), stateManager));
+        if (info.getLanguage() != null) {
+            map.put("language", messageSource.getMessage("language." + info.getLanguage().code(), locale));
+        } else {
+            map.put("language", messageSource.getMessage("search.err.language", locale));
+        }
+        map.put("ratingG", info.getRatingG());
+        map.put("ratingA", info.getRatingA());
+        map.put("lastMoveTime", info.getLastMoveTime() != null ? messageSource.formatDate(info.getLastMoveTime(), locale) : messageSource.getMessage("search.err.nomoves", locale));
+        map.put("activeGames", info.getActiveGames());
+        map.put("finishedGames", info.getFinishedGames());
+        map.put("playerOnline", stateManager.isPlayerOnline(Personality.person(info.getPid())));
+        map.put("averageMoveTime", info.getAverageMoveTime() != 0 ? messageSource.formatMinutes((long) (info.getAverageMoveTime() / 1000 / 60), locale) : messageSource.getMessage("search.err.nomoves", locale));
+    }
 
-	@Autowired
-	public void setStateManager(PlayerStateManager stateManager) {
-		this.stateManager = stateManager;
-	}
+    @Autowired
+    public void setStateManager(PlayerStateManager stateManager) {
+        this.stateManager = stateManager;
+    }
 
-	@Autowired
-	public void setMessageSource(GameMessageSource messageSource) {
-		this.messageSource = messageSource;
-	}
+    @Autowired
+    public void setMessageSource(GameMessageSource messageSource) {
+        this.messageSource = messageSource;
+    }
 
-	@Autowired
-	public void setPlayerSearchManager(ScribblePlayerSearchManager playerSearchManager) {
-		setEntitySearchManager(playerSearchManager);
-	}
+    @Autowired
+    public void setPlayerSearchManager(ScribblePlayerSearchManager playerSearchManager) {
+        setEntitySearchManager(playerSearchManager);
+    }
 }
